Тесты для задачи "идеальный вес" [#159458]

+++ src/test/java/ru/job4j/calculator/FitTest.java
@@ -9,7 +9,7 @@ public class FitTest {
     @Test
     public void WhenMan180Then92() {
         short in = 180;
-        int expected = 92;
+        double expected = 92;
         double out = Fit.manWeight(in);
         Assert.assertEquals(expected, out, 0.01);
     }
